requirements.txt
flask
ultralytics
opencv-python
pillow
numpy
python-dotenv
openai